moved append limits into own test fixture
one test was failing without this
