+++ src/EventStore.Grpc.Tests/Streams/append_to_stream.cs
@@ -5,9 +5,7 @@ using Xunit;
 
 namespace EventStore.Grpc.Streams {
 	[Trait("Category", "Network")]
-	public class append_to_stream :
-		IClassFixture<append_to_stream.Fixture> {
-		private const int MaxAppendSize = 1024;
+	public class append_to_stream : IClassFixture<append_to_stream.Fixture> {
 		private readonly Fixture _fixture;
 
 		public append_to_stream(Fixture fixture) {
@@ -15,8 +13,8 @@ namespace EventStore.Grpc.Streams {
 		}
 
 		public static IEnumerable<object[]> ExpectedVersionCreateStreamTestCases() {
-			yield return new object[] { AnyStreamRevision.Any, nameof(AnyStreamRevision.Any) };
-			yield return new object[] { AnyStreamRevision.NoStream, nameof(AnyStreamRevision.NoStream) };
+			yield return new object[] {AnyStreamRevision.Any, nameof(AnyStreamRevision.Any)};
+			yield return new object[] {AnyStreamRevision.NoStream, nameof(AnyStreamRevision.NoStream)};
 		}
 
 		[Theory, MemberData(nameof(ExpectedVersionCreateStreamTestCases))]
@@ -105,7 +103,7 @@ namespace EventStore.Grpc.Streams {
 			var stream = _fixture.GetStreamName();
 
 			var evnt = _fixture.CreateTestEvents().First();
-			var events = new []{evnt, evnt, evnt, evnt, evnt, evnt};
+			var events = new[] {evnt, evnt, evnt, evnt, evnt, evnt};
 
 			var writeResult = await _fixture.Client.AppendToStreamAsync(stream, AnyStreamRevision.Any, events);
 
@@ -338,49 +336,14 @@ namespace EventStore.Grpc.Streams {
 			await _fixture.Client.TombstoneAsync(stream, AnyStreamRevision.Any);
 
 			var result = await _fixture.Client.ConditionalAppendToStreamAsync(stream, AnyStreamRevision.Any,
-					_fixture.CreateTestEvents());
+				_fixture.CreateTestEvents());
 
 			Assert.Equal(ConditionalWriteResult.StreamDeleted, result);
 		}
 
-		[Fact]
-		public async Task succeeds_when_size_is_less_than_max_append_size() {
-			var stream = _fixture.GetStreamName();
-
-			await _fixture.Client.AppendToStreamAsync(
-				stream,
-				AnyStreamRevision.NoStream,
-				_fixture.GetEvents(MaxAppendSize - 1));
-		}
-
-		[Fact]
-		public async Task fails_when_size_exceeds_max_append_size() {
-			var stream = _fixture.GetStreamName();
-
-			await Assert.ThrowsAsync<MaximumAppendSizeExceededException>(() => _fixture.Client.AppendToStreamAsync(
-				stream,
-				AnyStreamRevision.NoStream,
-				_fixture.GetEvents(MaxAppendSize * 2)));
-		}
-
 		public class Fixture : EventStoreGrpcFixture {
-
-			public Fixture() : base(builder => builder.WithMaxAppendSize(MaxAppendSize)) {
-
-			}
 			protected override Task Given() => Task.CompletedTask;
 			protected override Task When() => Task.CompletedTask;
-
-			public IEnumerable<EventData> GetEvents(int maxSize) {
-				int size = 0;
-				foreach (var e in CreateTestEvents(int.MaxValue)) {
-					size += e.Data.Length;
-					if (size >= maxSize) {
-						yield break;
-					}
-					yield return e;
-				}
-			}
 		}
 	}
 }
